[Move] api -> plugin.api
プロジェクト構造上、ルートとなる `net.okocraft.box` にそのままコードを置くのは他モジュールと競合する原因となるので、ここからの開発は `net.okocraft.box.plugin` 下で行うものとする。
