+++ plugin/src/main/java/net/okocraft/box/Box.java
@@ -19,12 +19,12 @@
 package net.okocraft.box;
 
 import net.milkbowl.vault.economy.Economy;
-import net.okocraft.box.api.impl.BoxAPI;
 import net.okocraft.box.command.box.BoxCommand;
 import net.okocraft.box.command.boxadmin.BoxAdminCommand;
 import net.okocraft.box.listeners.BoxStick;
 import net.okocraft.box.listeners.PlayerListener;
 import net.okocraft.box.listeners.Replant;
+import net.okocraft.box.plugin.api.BoxAPI;
 import net.okocraft.box.util.APIRegisterer;
 import org.bukkit.Bukkit;
 import org.bukkit.plugin.Plugin;
