+++ plugin/src/main/java/net/okocraft/box/plugin/api/item/ApiItemManager.java
@@ -1,4 +1,4 @@
-package net.okocraft.box.api.impl.item;
+package net.okocraft.box.plugin.api.item;
 
 import net.okocraft.box.Box;
 import net.okocraft.box.api.item.BoxItem;
