+++ plugin/src/main/java/net/okocraft/box/plugin/api/user/ApiUserManager.java
@@ -1,4 +1,4 @@
-package net.okocraft.box.api.impl.user;
+package net.okocraft.box.plugin.api.user;
 
 import net.okocraft.box.api.user.User;
 import net.okocraft.box.api.user.UserManager;
